fix(frontend): extract type
